Refactor class names. no need for big table index.
