+++ app/src/main/java/com/games/ultimatetictactoe/app/Index.java
@@ -1,0 +1,10 @@
+package com.games.ultimatetictactoe.app;
+
+/**
+ * Created by alemjc on 11/5/15.
+ */
+public interface Index {
+     enum STATE {PLAYER1,PLAYER2,NONE};
+     STATE getState();
+     void setState(STATE state);
+}
